+++ lib/docopt/docopt_language_error.rb
@@ -1,0 +1,4 @@
+module Docopt
+  class DocoptLanguageError < SyntaxError
+  end
+end
